+++ apps/desktop/package.json
@@ -17,13 +17,16 @@
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@tailwindcss/typography": "^0.5.7",
     "@tauri-apps/cli": "^1.0.5",
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
     "@vitejs/plugin-react": "^2.0.0",
-    "autoprefixer": "^10.4.7",
-    "postcss": "^8.4.14",
+    "autoprefixer": "^10.4.8",
+    "postcss": "^8.4.16",
     "tailwind": "^4.0.0",
+    "tailwind-scrollbar-hide": "^1.1.7",
+    "tailwindcss": "^3.1.8",
     "typescript": "^4.7.4",
     "vite": "^3.0.3",
     "vite-plugin-tsconfig-paths": "^1.1.0"
