+++ apps/desktop/tailwind.config.js
@@ -1,0 +1,1 @@
+module.exports = require('../../common/config/tailwind.config.js');
