--- common/client/src/api/init.ts
@@ -1,3 +1,0 @@
-export async function startUpQuery() {
-	return '';
-}
