+++ common/interface/src/App.tsx
@@ -5,6 +5,7 @@ import {
 	queryClient,
 	useOnBoardingStore,
 } from '@stump/client';
+import { ChakraProvider } from '@chakra-ui/react';
 import { QueryClientProvider, defaultContext } from '@tanstack/react-query';
 import { ErrorBoundary } from 'react-error-boundary';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
@@ -12,6 +13,9 @@ import { ErrorFallback } from './components/ErrorFallback';
 import { useEffect, useState } from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import { AppRouter } from './AppRouter';
+import { chakraTheme } from './chakra';
+
+import './styles/index.css';
 
 function RouterContainer(props: { appProps: AppProps }) {
 	const [appProps, setAppProps] = useState(props.appProps);
@@ -40,13 +44,15 @@ function RouterContainer(props: { appProps: AppProps }) {
 
 export default function StumpInterface(props: AppProps) {
 	return (
-		<ErrorBoundary FallbackComponent={ErrorFallback}>
-			<QueryClientProvider client={queryClient} contextSharing={true}>
-				{import.meta.env.MODE === 'development' && (
-					<ReactQueryDevtools position="bottom-right" context={defaultContext} />
-				)}
-				<RouterContainer appProps={props} />
-			</QueryClientProvider>
-		</ErrorBoundary>
+		<ChakraProvider theme={chakraTheme}>
+			<ErrorBoundary FallbackComponent={ErrorFallback}>
+				<QueryClientProvider client={queryClient} contextSharing={true}>
+					{import.meta.env.MODE === 'development' && (
+						<ReactQueryDevtools position="bottom-right" context={defaultContext} />
+					)}
+					<RouterContainer appProps={props} />
+				</QueryClientProvider>
+			</ErrorBoundary>
+		</ChakraProvider>
 	);
 }
