+++ common/interface/src/AppLayout.tsx
@@ -1,17 +1,27 @@
-import { Box, Flex, useColorModeValue } from '@chakra-ui/react';
-import { AppPropsContext, useStartUpQuery } from '@stump/client';
 import React, { useContext } from 'react';
 import { Outlet } from 'react-router-dom';
+
+import { Box, Flex, useColorModeValue } from '@chakra-ui/react';
+import { AppPropsContext } from '@stump/client';
+
 import Lazy from './components/Lazy';
-// import Sidebar from './components/Sidebar/Sidebar';
 
 export function AppLayout() {
 	const appProps = useContext(AppPropsContext);
 
-	const { data } = useStartUpQuery();
-
 	return (
-		<Flex w="full" h="full">
+		<Flex
+			w="full"
+			h="full"
+			onContextMenu={(e) => {
+				if (appProps?.platform != 'browser') {
+					e.preventDefault();
+					return false;
+				}
+
+				return true;
+			}}
+		>
 			{/* <Sidebar /> */}
 			{/* {!hideSidebar && <Sidebar />} */}
 			<Box as="main" w="full" h="full" bg={useColorModeValue('gray.75', 'gray.900')}>
